plugin/dnstap: various cleanups
A recent issue made me look into this plugin, I suspect various other
cleanups (hopefully deletion of code) can be made as well

Remove identical functions ToClientQuery etc, and just use tap.Message
as the base type in plugin. This removes the msg/ directory.

This remove the whole test directory as we will just check the things we
are interested in which gives much better feedback.

tapwr dir is also not needed, writer_test.go was just duplicating the
tests already done. This moves writer.go to the top directory.

Make the only user of dnstap, the forward plugin, use the newer code
also remove the test, a better test there would be a full e2e test to
see the correct thing happens.

Signed-off-by: Miek Gieben <miek@miek.nl>

typo

Signed-off-by: Miek Gieben <miek@miek.nl>

Any test here should be a e2e test, not a shallow retest of the dnstap pkg

Signed-off-by: Miek Gieben <miek@miek.nl>